[types] Migrate MoveString & MoveAsciiString to move_std
