+++ crates/rooch-types/src/framework/auth_validator.rs
@@ -17,7 +17,7 @@ use moveos_types::move_option::MoveOption;
 use moveos_types::transaction::MoveAction;
 use moveos_types::{
     module_binding::MoveFunctionCaller,
-    move_string::MoveAsciiString,
+    move_std::ascii::MoveAsciiString,
     move_types::FunctionId,
     moveos_std::tx_context::TxContext,
     state::{MoveStructState, MoveStructType},
