+++ moveos/moveos-types/src/move_std/mod.rs
@@ -1,5 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod ascii;
 /// This mod contains all the Rust to Move type mapping that are used in MoveStd
 pub mod error;
+pub mod string;
