+++ moveos/moveos-types/src/moveos_std/tx_context.rs
@@ -8,7 +8,7 @@ use crate::gas_config::GasConfig;
 use crate::h256::{self, H256};
 use crate::move_any::{AnyTrait, CopyableAny};
 use crate::move_simple_map::SimpleMap;
-use crate::move_string::MoveString;
+use crate::move_std::string::MoveString;
 use crate::moveos_std::object::ObjectID;
 use crate::state::{MoveState, MoveStructState, MoveStructType};
 use anyhow::Result;
