+++ moveos/moveos-types/src/moveos_std/tx_meta.rs
@@ -4,7 +4,7 @@
 use crate::{
     addresses::MOVEOS_STD_ADDRESS,
     move_option::MoveOption,
-    move_string::MoveAsciiString,
+    move_std::ascii::MoveAsciiString,
     state::{MoveStructState, MoveStructType},
     transaction::MoveAction,
 };
